Add new texture for slots
Slots are not working properly yet because the `CaseScreenHandler` is seemingly created twice per GUI opening. Once from `CaseEntity` and once from its registration in`Components.kt`. The latter is generic across Case tiers, and so right now passes in 0 for the tier, which causes an array index out of bounds error.